Add code-base, refactor
