+++ addon/components/code-block.hbs
@@ -3,6 +3,7 @@
 {{/unless}}
 
 <pre
+  ...attributes
   class="code-block {{this.languageClass}} {{if @showLineNumbers "line-numbers"}}"
 >
   {{~! ~}}
