+++ addon/components/code-inline.hbs
@@ -3,6 +3,7 @@
 {{/unless}}
 
 <code
+  ...attributes
   class="code-inline {{this.languageClass}}"
   {{did-insert this.setPrismCode}}
   {{did-update this.setPrismCode this.code @language}}
