--- addon/templates/components/code-block.hbs
@@ -1,1 +1,0 @@
-{{#if hasBlock}}{{#in-element this.blockElement}}{{yield}}{{/in-element}}{{/if}}<code ...attributes class={{this.languageClass}}>{{this.prismCode}}</code>
