+++ app/components/code-block.js
@@ -1,3 +1,1 @@
-import CodeBlock from 'ember-prism/components/code-block';
-
-export default CodeBlock;
+export { default } from 'ember-prism/components/code-block';
