--- app/components/new-code-inline.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-prism/components/new-code-inline';
